debbuging many console errors
