+++ .babelrc
@@ -1,18 +1,19 @@
 {
   "presets": [
-    ["env", {
+    ["@babel/preset-env", {
       "modules": false,
       "targets": {
         "browsers": "> 1%",
         "uglify": true
       },
-      "useBuiltIns": true
+      "useBuiltIns": "usage",
+      "corejs": 3
     }]
   ],
-
   "plugins": [
-    "syntax-dynamic-import",
-    "transform-object-rest-spread",
-    ["transform-class-properties", { "spec": true }]
+    "@babel/plugin-syntax-dynamic-import",
+    "@babel/plugin-proposal-optional-chaining", // Add this line
+    "@babel/plugin-transform-object-rest-spread",
+    ["@babel/plugin-transform-class-properties", { "spec": true }]
   ]
 }
