+++ app/views/layouts/application.html.erb
@@ -6,10 +6,7 @@
   <%= csrf_meta_tags %>
   <%= csp_meta_tag %>
   <%= stylesheet_link_tag "application", "data-turbo-track": "reload" %>
-  <%= javascript_include_tag "application", "data-turbo-track": "reload" %>
-  <%= javascript_importmap_tags %>
-  <!-- Add the script tag for your ES module -->
-  <script type="module" src="js/index.js"></script>
+  <%= javascript_pack_tag "application", "data-turbo-track": "reload" %>
 </head>
 <body>
   <%= render "shared/flashes" %>
