+++ package.json
@@ -1,10 +1,21 @@
 {
   "dependencies": {
+    "@hotwired/turbo-rails": "^8.0.4",
     "@rails/webpacker": "5.4.4",
+    "babel-preset-env": "^1.7.0",
+    "bootstrap": "^5.3.3",
     "webpack": "^4.46.0",
     "webpack-cli": "^3.3.12"
   },
   "devDependencies": {
+    "@babel/core": "^7.24.5",
+    "@babel/plugin-proposal-optional-chaining": "^7.21.0",
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "@babel/plugin-transform-private-methods": "^7.24.1",
+    "@babel/preset-env": "^7.24.5",
+    "babel-plugin-syntax-dynamic-import": "^6.18.0",
+    "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-plugin-transform-object-rest-spread": "^6.26.0",
     "webpack-dev-server": "^3"
   }
 }
